Big refactoring, bug fixes and push functions
No relaying all the oncollisionenter of the rigid bodies of the
children in the parent via message upward. Also using collision
detection continuous instead of discrete in order to avoid
interpenetration. removed the scripts in children parts, let’s try to
keep all the scripting parts on the parent object.